# doc/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TAO Test Center API",
    "version": "0.1"
  },
  "paths": {
    "/taoTestCenterRostering/api/eligibilities": {
      "get": {
        "tags": [
          "eligibilities"
        ],
        "summary": "Search for eligibilities",
        "description": "Search for eligibilities",
        "operationId": "oat\\taoTestCenter\\controller\\RestEligibilities::get",
        "parameters": [
          {
            "name": "delivery",
            "in": "query",
            "description": "Delivery Uri (Url encoded)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "testCenter",
            "in": "query",
            "description": "Test center Uri (Url encoded)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Eligibility data",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "description": "`false` on failure, `true` on success",
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Eligibility"
                      }
                    }
                  },
                  "type": "object",
                  "example": {
                    "success": true,
                    "data": [
                      {
                        "delivery": "http://sample/first.rdf#i1536680377163170",
                        "testCenter": "http://sample/first.rdf#i1536680377163171",
                        "testTakers": [
                          "http://sample/first.rdf#i1536680377163172",
                          "http://sample/first.rdf#i1536680377163173"
                        ]
                      }
                    ]
                  }
                }
              }
            }
          },
          "404": {
            "description": "Eligibility not found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": false,
                    "errorCode": 404,
                    "errorMsg": "Eligibility not found for provided search parameters.",
                    "version": "3.3.0-sprint85"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Eligibility": {
        "required": [
          "delivery",
          "testCenter"
        ],
        "properties": {
          "delivery": {
            "description": "delivery URI",
            "type": "string"
          },
          "testTakers": {
            "description": "Array of test-taker URIs",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "testCenter": {
            "description": "Test center URI",
            "type": "string"
          },
          "id": {
            "description": "Eligibility identifier",
            "type": "string"
          },
          "proctored": {
            "description": "Eligibility proctored   value",
            "type": "boolean"
          }
        },
        "type": "object"
      }
    }
  }
}